# Call commit callbacks from the tail of the list