@lidongyang 2017-12-21
@behlendorf yep you are right about tgt_cb_last_committed(), and it did improve performance numbers, we got them using mdtest:
Before the patch:
-- started at 12/15/2017 18:23:34 --

mdtest-1.9.4-rc1 was launched with 120 total task(s) on 30 node(s)
Command line used: /system/Benchmarks/Filesystem/mdtest/mdtest_ompi211 -u -d /g/data1b/fu2/jas900/mdtest/mdt0/25@/g/data1b/fu2/jas900/mdtest/mdt1/25@/g/data1b/fu2/jas900/mdtest/mdt2/25@/g/data1b/fu2/jas900/mdtest/mdt3/25 -n 32768 -i 2
Path: /g/data1b/fu2/jas900/mdtest/mdt0
FS: 11714.4 TiB   Used FS: 0.1%   Inodes: 191.8 Mi   Used Inodes: 24.9%

120 tasks, 3932160 files/directories

SUMMARY: (of 2 iterations)
   Operation                      Max            Min           Mean        Std Dev
   ---------                      ---            ---           ----        -------
   Directory creation:      46209.655      45325.344      45767.499        442.155
   Directory stat    :     422943.103     395433.941     409188.522      13754.581
   Directory removal :     109555.447     107923.481     108739.464        815.983
   File creation     :      84488.802      82024.716      83256.759       1232.043
   File stat         :     290506.847     260661.397     275584.122      14922.725
   File read         :     308452.944     306049.686     307251.315       1201.629
   File removal      :     125729.192     122936.462     124332.827       1396.365
   Tree creation     :        280.482        262.247        271.365          9.118
   Tree removal      :         56.103         28.919         42.511         13.592

-- finished at 12/15/2017 18:31:30 --
Note that mdt will get soft lockup and eventually crash.

After the patch:
-- started at 12/16/2017 18:22:51 --

mdtest-1.9.4-rc1 was launched with 120 total task(s) on 30 node(s)
Command line used: /system/Benchmarks/Filesystem/mdtest/mdtest_ompi211 -u -d /g/data1b/fu2/jas900/mdtest/mdt0/26@/g/data1b/fu2/jas900/mdtest/mdt1/26@/g/data1b/fu2/jas900/mdtest/mdt2/26@/g/data1b/fu2/jas900/mdtest/mdt3/26 -n 32768 -i 2
Path: /g/data1b/fu2/jas900/mdtest/mdt0
FS: 11714.4 TiB   Used FS: 0.1%   Inodes: 191.6 Mi   Used Inodes: 25.6%

120 tasks, 3932160 files/directories

SUMMARY: (of 2 iterations)
   Operation                      Max            Min           Mean        Std Dev
   ---------                      ---            ---           ----        -------
   Directory creation:      54567.094      53627.233      54097.164        469.930
   Directory stat    :     430300.347     389450.366     409875.357      20424.991
   Directory removal :     155325.535     155181.753     155253.644         71.891
   File creation     :     129396.860     125304.163     127350.511       2046.348
   File stat         :     305263.708     301667.676     303465.692       1798.016
   File read         :     305201.679     303282.957     304242.318        959.361
   File removal      :     199668.640     198088.652     198878.646        789.994
   Tree creation     :        677.521        371.513        524.517        153.004
   Tree removal      :         51.097         31.695         41.396          9.701

-- finished at 12/16/2017 18:28:59 --

For the operations which will generate a transaction and a new last_commited, like creation and removal we can see a improvement. for others like stat and read, nothing really changed. It makes sense since the patch only changes trans callbacks.

Thanks
Dongyang